# docs/swagger.yaml
basePath: /
definitions:
  system.Button:
    properties:
      code:
        type: string
      desc:
        type: string
      id:
        type: integer
    type: object
  system.Menu:
    properties:
      activeMenu:
        description: 激活的菜单名
        type: string
      buttons:
        items:
          $ref: '#/definitions/system.Button'
        type: array
      children:
        items:
          $ref: '#/definitions/system.Menu'
        type: array
      component:
        type: string
      constant:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        type: string
      fixedIndexInTab:
        type: integer
      hideInMenu:
        description: 是否隐藏菜单
        type: boolean
      href:
        type: string
      i18nKey:
        type: string
      icon:
        type: string
      iconType:
        type: string
      id:
        type: integer
      keepAlive:
        type: boolean
      localIcon:
        type: string
      menuName:
        type: string
      menuType:
        type: string
      multiTab:
        description: 是否支持多标签
        type: boolean
      order:
        type: integer
      parentId:
        description: 父菜单ID
        type: integer
      query:
        items:
          $ref: '#/definitions/system.MenuQuery'
        type: array
      routeName:
        type: string
      routePath:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  system.MenuQuery:
    properties:
      id:
        type: integer
      key:
        type: string
      menuId:
        type: integer
      value:
        type: string
    type: object
...
  utils.Response-array_system_Button:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/system.Button'
        type: array
      msg:
        type: string
    type: object
  utils.Response-array_system_Menu:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/system.Menu'
        type: array
      msg:
        type: string
    type: object
...
  utils.Response-string:
    properties:
      code:
        type: string
      data:
        type: string
      msg:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Soybean Admin Go 后台管理系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Soybean Admin Go API
  version: "1.0"
paths:
...
  /buttons:
    get:
      consumes:
      - application/json
      description: 获取系统所有按钮列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Button'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取所有按钮
      tags:
      - 按钮管理
  /buttons/role:
    get:
      consumes:
      - application/json
      description: 获取指定角色的按钮列表
      parameters:
      - description: 角色ID
        in: query
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Button'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取角色按钮
      tags:
      - 按钮管理
  /buttons/user:
    get:
      consumes:
      - application/json
      description: 获取当前用户的按钮权限
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Button'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取用户按钮
      tags:
      - 按钮管理
  /menus:
    get:
      consumes:
      - application/json
      description: 获取系统所有菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Menu'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取所有菜单
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建新菜单
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/system.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Menu'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /menus/batchDelete:
    post:
      consumes:
      - application/json
      description: 批量删除多个菜单
      parameters:
      - description: 菜单ID列表
        in: body
        name: ids
        required: true
        schema:
          properties:
            ids:
              items:
                type: integer
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-string'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 批量删除菜单
      tags:
      - 菜单管理
  /menus/delete:
    post:
      consumes:
      - application/json
      description: 删除指定菜单
      parameters:
      - description: 菜单ID
        in: body
        name: id
        required: true
        schema:
          properties:
            id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-string'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
  /menus/role:
    get:
      consumes:
      - application/json
      description: 获取指定角色的菜单列表
      parameters:
      - description: 角色ID
        in: query
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Menu'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取角色菜单
      tags:
      - 菜单管理
  /menus/update:
    post:
      consumes:
      - application/json
      description: 更新菜单信息
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/system.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Menu'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 更新菜单
      tags:
      - 菜单管理
  /menus/user:
    get:
      consumes:
      - application/json
      description: 获取当前用户的菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Menu'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取用户菜单
      tags:
      - 菜单管理
...
securityDefinitions:
  ApiKeyAuth:
    description: '请在此处输入Bearer Token，格式为: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"